requirements.txt
boto3==1.34.68
Django==4.2.11
django-storages==1.14.2
djangorestframework==3.15.1
python-dotenv==1.0.1
requests==2.31.0
#twelvelabs==0.1.21
huey==2.5.0
huey-django-orm==23.12.10.0
pytransloadit==0.2.2
